// src/modules/post/post.interface.ts
import { PostStatus } from "../../../generated/prisma/enums";

export interface ICreatePostPayload {
    title: string;
    content: string;
    thumbnail: string;
    isFeatured?: boolean;
    status?: PostStatus
    tags: string[]
}


export interface IUpdatePostload {
    title?: string;
    content?: string;
    thumbnail?: string;
    isFeatured?: boolean;
    status?: PostStatus
    tags?: string[]
}